#68 built a graphql API url component
